--- src/main/java/com/github/pitzzahh/utilities/classes/enums/Gender.java
@@ -1,23 +1,0 @@
-package com.github.pitzzahh.utilities.classes.enums;
-
-/**
- * enum template for genders.
- */
-public enum Gender {
-    /**
-     * Select this if you are a male
-     */
-    MALE,
-    /**
-     * Select this if you are a female
-     */
-    FEMALE,
-    /**
-     * Select this if you are a part of lgbt
-     */
-    LGBT,
-    /**
-     * Select this if you are preferred not to say your gender.
-     */
-    PREFER_NOT_TO_SAY
-}
