--- src/test/java/com/github/pitzzahh/utilities/classes/PersonTest.java
@@ -1,12 +1,0 @@
-package com.github.pitzzahh.utilities.classes;
-
-import org.junit.jupiter.api.Test;
-
-class PersonTest {
-
-    @Test
-    void canCreateEmptyPerson() {
-        Person person = Person.builder().build();
-        System.out.println("person = " + person);
-    }
-}
